requirements.txt
huggingface_hub==0.29.1
datasets>=3.3.2
pandas>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.1
aiohttp>=3.8.5
asyncio>=3.4.3
transformers==4.49.0